Moved NodeBuilderDescriptor to its own file
