+++ lluvia/cpp/core/include/lluvia/core.h
@@ -38,6 +38,7 @@
 #include "core/node/PortDirection.h"
 #include "core/node/PortType.h"
 #include "core/node/PushConstants.h"
+#include "core/node/NodeBuilderDescriptor.h"
 
 #include "core/CommandBuffer.h"
 #include "core/ContainerNode.h"
@@ -45,7 +46,6 @@
 #include "core/Duration.h"
 #include "core/error.h"
 #include "core/Interpreter.h"
-#include "core/NodeBuilderDescriptor.h"
 #include "core/Parameter.h"
 #include "core/Program.h"
 #include "core/Session.h"
