+++ lluvia/python/src/lluvia/core/session.pxd
@@ -23,9 +23,10 @@ from lluvia.core.duration cimport _Duration
 
 from lluvia.core.node_new.compute_node cimport _ComputeNode
 from lluvia.core.node_new.compute_node_descriptor cimport _ComputeNodeDescriptor
+from lluvia.core.node_new.node_builder_descriptor cimport _NodeBuilderDescriptor
 
 from lluvia.core.device.device_descriptor cimport _DeviceDescriptor
-from lluvia.core.node cimport _ContainerNodeDescriptor, _ContainerNode, _NodeBuilderDescriptor
+from lluvia.core.node cimport _ContainerNodeDescriptor, _ContainerNode
 from lluvia.core.program cimport _Program
 from lluvia.core.types cimport _vec3ui
