+++ lluvia/python/src/lluvia/core/session.pyx
@@ -48,11 +48,11 @@ from lluvia.core.program cimport Program
 
 from lluvia.core.node cimport ContainerNodeDescriptor,\
                               ContainerNode,\
-                              NodeBuilderDescriptor,\
                               _buildContainerNode
 
 from lluvia.core.node_new.compute_node cimport ComputeNode, _buildComputeNode
 from lluvia.core.node_new.compute_node_descriptor cimport ComputeNodeDescriptor
+from lluvia.core.node_new.node_builder_descriptor cimport NodeBuilderDescriptor
 
 from lluvia.core.types cimport _vec3ui
